+++ src/collective/contact/plonegroup/subscribers.py
@@ -1,20 +1,24 @@
 # -*- coding: utf-8 -*-
 from Acquisition import aq_get
-from zope.component import getUtility
+from zExceptions import Redirect
+from zope.component import getUtility, getMultiAdapter
 from zope.interface import alsoProvides, Interface, noLongerProvides
 from zope.lifecycleevent.interfaces import IObjectRemovedEvent
 from zope.schema import getFieldsInOrder
 from zope.schema.interfaces import ICollection, IText, IChoice
 
 from plone import api
-from plone.behavior.interfaces import IBehavior
-from plone.dexterity.interfaces import IDexterityContent, IDexterityFTI
 from plone.app.linkintegrity.interfaces import ILinkIntegrityInfo
 from plone.app.linkintegrity.handlers import referencedObjectRemoved as baseReferencedObjectRemoved
+from plone.behavior.interfaces import IBehavior
+from plone.dexterity.interfaces import IDexterityContent, IDexterityFTI
+from plone.registry.interfaces import IRegistry
 
 from Products.CMFPlone.utils import base_hasattr
+from Products.statusmessages.interfaces import IStatusMessage
 
-from config import PLONEGROUP_ORG
+from . import _
+from config import PLONEGROUP_ORG, ORGANIZATIONS_REGISTRY
 from interfaces import IPloneGroupContact, INotPloneGroupContact
 
 try:
@@ -114,6 +118,34 @@ def plonegroupOrganizationRemoved(del_obj, event):
 def referencedObjectRemoved(obj, event):
     if not IReferenceable.providedBy(obj):
         baseReferencedObjectRemoved(obj, event)
+
+
+def plonegroup_contact_transition(contact, event):
+    """
+        React when a IPloneGroupContact transition is done
+    """
+    if event.transition.id == 'deactivate':
+        # check if the transition is selected
+        registry = getUtility(IRegistry)
+        errors = []
+        if contact.UID() in registry[ORGANIZATIONS_REGISTRY]:
+            errors.append(_('This contact is selected in configuration'))
+        else:
+            search_value_in_objects(contact, contact.UID(), p_types=[], type_fields={})
+            storage = ILinkIntegrityInfo(contact.REQUEST)
+            breaches = storage.getIntegrityBreaches()
+            if contact in breaches:
+                errors.append(_("This contact is used in following content: ${items}",
+                                mapping={'items': ', '.join(['<a href="%s" target="_blank">%s</a>'
+                                                             % (i.absolute_url(), i.Title())
+                                                             for i in breaches[contact]])}))
+        if errors:
+            smi = IStatusMessage(contact.REQUEST)
+            smi.addStatusMessage(_('You cannot deactivate this item !'), type='error')
+            smi.addStatusMessage(errors[0], type='error')
+            view_url = getMultiAdapter((contact, contact.REQUEST), name=u'plone_context_state').view_url()
+            #contact.REQUEST['RESPONSE'].redirect(view_url)
+            raise Redirect(view_url)
 
 
 def mark_organization(contact, event):